PROD-341 PostContent 렌더러를 분리한다
