+++ apps/app/src/components/post/PostBody.tsx
@@ -1,8 +1,5 @@
-import { isPostContentDocumentV1 } from '@kosmo/core/post-content';
-import { Linking, StyleSheet, Text, View } from 'react-native';
 import { graphql, useFragment } from 'react-relay';
-import { useTheme } from '@/theme/ThemeProvider';
-import { typography } from '@/theme/tokens';
+import { PostContentRenderer } from './PostContentRenderer';
 import type { PostBody_post$key } from './__generated__/PostBody_post.graphql';
 
 const PostBodyFragment = graphql`
@@ -23,7 +20,6 @@ export function PostBody({
   post: PostBody_post$key;
   size?: 'md' | 'lg';
 }) {
-  const theme = useTheme();
   const post = useFragment(PostBodyFragment, postKey);
   const content = post.content;
 
@@ -31,60 +27,7 @@ export function PostBody({
     return null;
   }
 
-  const document = isPostContentDocumentV1(content.document) ? content.document.body : null;
-
-  if (!document) {
-    return content.bodyText ? (
-      <Text
-        style={[styles.body, size === 'lg' ? typography.lg : typography.md, { color: theme.text }]}
-      >
-        {content.bodyText}
-      </Text>
-    ) : null;
-  }
-
   return (
-    <View>
-      {document.content.map((paragraph, paragraphIndex) => (
-        <Text
-          key={paragraphIndex}
-          style={[
-            styles.body,
-            size === 'lg' ? typography.lg : typography.md,
-            { color: theme.text },
-          ]}
-        >
-          {(paragraph.content ?? []).map((node, inlineIndex) => {
-            if (node.type === 'hard_break') {
-              return '\n';
-            }
-            const link = node.marks?.[0];
-
-            return link ? (
-              <Text
-                accessibilityLabel={`${node.text}, ${link.attrs.href}`}
-                accessibilityRole="link"
-                key={inlineIndex}
-                onPress={(event) => {
-                  event.preventDefault();
-                  event.stopPropagation();
-                  void Linking.openURL(link.attrs.href);
-                }}
-                style={styles.link}
-              >
-                {node.text}
-              </Text>
-            ) : (
-              node.text
-            );
-          })}
-        </Text>
-      ))}
-    </View>
+    <PostContentRenderer bodyText={content.bodyText} document={content.document} size={size} />
   );
 }
-
-const styles = StyleSheet.create({
-  body: { fontFamily: 'Pretendard' },
-  link: { textDecorationLine: 'underline' },
-});
